Added basic thread safety for OpenMP applications by encapsulating all logic into MASTER pragmas.  Only the performance of the master thread is measured.
Now only the log format uses the raw hash numbers for enumerating the
profiles. The tree-format uses a simple numbering from 0 to N-1, where N
is the number of process froups with distinct timers. Indexing is
according to the number of processes, the 0th profile will have the
profile for the one with the most processes.